+++ src/models/pdearena/unet.py
@@ -1,65 +1,21 @@
-from typing import List, Tuple, Union
+r"""
+Taken and modified from pdearena
+"""
+
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT license.
+from typing import List, Optional, Tuple, Union
 
 import torch
-import torch.nn as nn
+from torch import nn
 
-class SpectralConv2d(nn.Module):
-    """2D Fourier layer. Does FFT, linear transform, and Inverse FFT.
-    Implemented in a way to allow multi-gpu training.
-    Args:
-        in_channels (int): Number of input channels
-        out_channels (int): Number of output channels
-        modes1 (int): Number of Fourier modes to keep in the first spatial direction
-        modes2 (int): Number of Fourier modes to keep in the second spatial direction
-    [paper](https://arxiv.org/abs/2010.08895)
-    """
+from .activations import ACTIVATION_REGISTRY
+from .fourier import SpectralConv2d
 
-    def __init__(self, in_channels: int, out_channels: int, modes1: int, modes2: int):
-        super().__init__()
+# Largely based on https://github.com/labmlai/annotated_deep_learning_paper_implementations/blob/master/labml_nn/diffusion/ddpm/unet.py
+# MIT License
+# Copyright (c) 2020 Varuna Jayasiri
 
-        self.in_channels = in_channels
-        self.out_channels = out_channels
-        self.modes1 = modes1  # Number of Fourier modes to multiply, at most floor(N/2) + 1
-        self.modes2 = modes2
-
-        self.scale = 1 / (in_channels * out_channels)
-        self.weights1 = nn.Parameter(
-            self.scale * torch.rand(in_channels, out_channels, self.modes1, self.modes2, 2, dtype=torch.float32)
-        )
-        self.weights2 = nn.Parameter(
-            self.scale * torch.rand(in_channels, out_channels, self.modes1, self.modes2, 2, dtype=torch.float32)
-        )
-
-    def forward(self, x, x_dim=None, y_dim=None):
-        batchsize = x.shape[0]
-        # Compute Fourier coeffcients up to factor of e^(- something constant)
-        x_ft = torch.fft.rfft2(x)
-
-        # Multiply relevant Fourier modes
-        out_ft = torch.zeros(
-            batchsize,
-            self.out_channels,
-            x.size(-2),
-            x.size(-1) // 2 + 1,
-            dtype=torch.cfloat,
-            device=x.device,
-        )
-        out_ft[:, :, : self.modes1, : self.modes2] = SpectralConv2d._batchmul2d(
-            x_ft[:, :, : self.modes1, : self.modes2], torch.view_as_complex(self.weights1)
-        )
-        out_ft[:, :, -self.modes1 :, : self.modes2] = SpectralConv2d._batchmul2d(
-            x_ft[:, :, -self.modes1 :, : self.modes2], torch.view_as_complex(self.weights2)
-        )
-
-        # Return to physical space
-        x = torch.fft.irfft2(out_ft, s=(x.size(-2), x.size(-1)))
-        return x
-    
-    @staticmethod
-    def _batchmul2d(input, weights):
-        # (batch, in_channel, x,y ), (in_channel, out_channel, x,y) -> (batch, out_channel, x,y)
-        return torch.einsum("bixy,ioxy->boxy", input, weights)
-    
 
 class ResidualBlock(nn.Module):
     """Wide Residual Blocks used in modern Unet architectures.
@@ -67,6 +23,7 @@ class ResidualBlock(nn.Module):
     Args:
         in_channels (int): Number of input channels.
         out_channels (int): Number of output channels.
+        activation (str): Activation function to use.
         norm (bool): Whether to use normalization.
         n_groups (int): Number of groups for group normalization.
     """
@@ -75,11 +32,14 @@ class ResidualBlock(nn.Module):
         self,
         in_channels: int,
         out_channels: int,
+        activation: str = "gelu",
         norm: bool = False,
         n_groups: int = 1,
     ):
         super().__init__()
-
+        self.activation: nn.Module = ACTIVATION_REGISTRY.get(activation, None)
+        if self.activation is None:
+            raise NotImplementedError(f"Activation {activation} not implemented")
         self.conv1 = nn.Conv2d(in_channels, out_channels, kernel_size=(3, 3), padding=(1, 1))
         self.conv2 = nn.Conv2d(out_channels, out_channels, kernel_size=(3, 3), padding=(1, 1))
         # If the number of input channels is not equal to the number of output channels we have to
@@ -95,7 +55,6 @@ class ResidualBlock(nn.Module):
         else:
             self.norm1 = nn.Identity()
             self.norm2 = nn.Identity()
-        self.activation = nn.GELU()
 
     def forward(self, x: torch.Tensor):
         # First convolution layer
@@ -113,7 +72,8 @@ class FourierResidualBlock(nn.Module):
         in_channels (int): Number of input channels.
         out_channels (int): Number of output channels.
         modes1 (int): Number of modes in the first dimension.
-        modes2 (int): Number of modes in the second dimension..
+        modes2 (int): Number of modes in the second dimension.
+        activation (str): Activation function to use.
         norm (bool): Whether to use normalization.
         n_groups (int): Number of groups for group normalization.
     """
@@ -124,10 +84,14 @@ class FourierResidualBlock(nn.Module):
         out_channels: int,
         modes1: int = 16,
         modes2: int = 16,
+        activation: str = "gelu",
         norm: bool = False,
         n_groups: int = 1,
     ):
         super().__init__()
+        self.activation: nn.Module = ACTIVATION_REGISTRY.get(activation, None)
+        if self.activation is None:
+            raise NotImplementedError(f"Activation {activation} not implemented")
 
         self.modes1 = modes1
         self.modes2 = modes2
@@ -149,7 +113,6 @@ class FourierResidualBlock(nn.Module):
         else:
             self.norm1 = nn.Identity()
             self.norm2 = nn.Identity()
-        self.activation = nn.GELU()
 
     def forward(self, x: torch.Tensor):
         # using pre-norms
@@ -164,6 +127,64 @@ class FourierResidualBlock(nn.Module):
         return out
 
 
+class AttentionBlock(nn.Module):
+    """Attention block This is similar to [transformer multi-head
+    attention]
+
+    Args:
+        n_channels (int): the number of channels in the input
+        n_heads (int): the number of heads in multi-head attention
+        d_k: the number of dimensions in each head
+        n_groups (int): the number of groups for [group normalization][torch.nn.GroupNorm].
+
+    """
+
+    def __init__(self, n_channels: int, n_heads: int = 1, d_k: Optional[int] = None, n_groups: int = 1):
+        super().__init__()
+
+        # Default `d_k`
+        if d_k is None:
+            d_k = n_channels
+        # Normalization layer
+        self.norm = nn.GroupNorm(n_groups, n_channels)
+        # Projections for query, key and values
+        self.projection = nn.Linear(n_channels, n_heads * d_k * 3)
+        # Linear layer for final transformation
+        self.output = nn.Linear(n_heads * d_k, n_channels)
+        # Scale for dot-product attention
+        self.scale = d_k**-0.5
+        #
+        self.n_heads = n_heads
+        self.d_k = d_k
+
+    def forward(self, x: torch.Tensor):
+        # Get shape
+        batch_size, n_channels, height, width = x.shape
+        # Change `x` to shape `[batch_size, seq, n_channels]`
+        x = x.view(batch_size, n_channels, -1).permute(0, 2, 1)
+        # Get query, key, and values (concatenated) and shape it to `[batch_size, seq, n_heads, 3 * d_k]`
+        qkv = self.projection(x).view(batch_size, -1, self.n_heads, 3 * self.d_k)
+        # Split query, key, and values. Each of them will have shape `[batch_size, seq, n_heads, d_k]`
+        q, k, v = torch.chunk(qkv, 3, dim=-1)
+        # Calculate scaled dot-product $\frac{Q K^\top}{\sqrt{d_k}}$
+        attn = torch.einsum("bihd,bjhd->bijh", q, k) * self.scale
+        # Softmax along the sequence dimension $\underset{seq}{softmax}\Bigg(\frac{Q K^\top}{\sqrt{d_k}}\Bigg)$
+        attn = attn.softmax(dim=1)
+        # Multiply by values
+        res = torch.einsum("bijh,bjhd->bihd", attn, v)
+        # Reshape to `[batch_size, seq, n_heads * d_k]`
+        res = res.view(batch_size, -1, self.n_heads * self.d_k)
+        # Transform to `[batch_size, seq, n_channels]`
+        res = self.output(res)
+
+        # Add skip connection
+        res += x
+
+        # Change to shape `[batch_size, in_channels, height, width]`
+        res = res.permute(0, 2, 1).view(batch_size, n_channels, height, width)
+        return res
+
+
 class DownBlock(nn.Module):
     """Down block This combines [`ResidualBlock`][pdearena.modules.twod_unet.ResidualBlock] and [`AttentionBlock`][pdearena.modules.twod_unet.AttentionBlock].
 
@@ -172,6 +193,8 @@ class DownBlock(nn.Module):
     Args:
         in_channels (int): Number of input channels
         out_channels (int): Number of output channels
+        has_attn (bool): Whether to use attention block
+        activation (nn.Module): Activation function
         norm (bool): Whether to use normalization
     """
 
@@ -179,13 +202,21 @@ class DownBlock(nn.Module):
         self,
         in_channels: int,
         out_channels: int,
+        has_attn: bool = False,
+        activation: str = "gelu",
         norm: bool = False,
     ):
         super().__init__()
-        self.res = ResidualBlock(in_channels, out_channels, norm=norm)
+        self.res = ResidualBlock(in_channels, out_channels, activation=activation, norm=norm)
+        if has_attn:
+            self.attn = AttentionBlock(out_channels)
+        else:
+            self.attn = nn.Identity()
 
     def forward(self, x: torch.Tensor):
-        return self.res(x)
+        x = self.res(x)
+        x = self.attn(x)
+        return x
 
 
 class FourierDownBlock(nn.Module):
@@ -200,6 +231,8 @@ class FourierDownBlock(nn.Module):
         out_channels: int,
         modes1: int = 16,
         modes2: int = 16,
+        has_attn: bool = False,
+        activation: str = "gelu",
         norm: bool = False,
     ):
         super().__init__()
@@ -208,11 +241,18 @@ class FourierDownBlock(nn.Module):
             out_channels,
             modes1=modes1,
             modes2=modes2,
+            activation=activation,
             norm=norm,
         )
+        if has_attn:
+            self.attn = AttentionBlock(out_channels)
+        else:
+            self.attn = nn.Identity()
 
     def forward(self, x: torch.Tensor):
-        return self.res(x)
+        x = self.res(x)
+        x = self.attn(x)
+        return x
 
 
 class UpBlock(nn.Module):
@@ -223,6 +263,8 @@ class UpBlock(nn.Module):
     Args:
         in_channels (int): Number of input channels
         out_channels (int): Number of output channels
+        has_attn (bool): Whether to use attention block
+        activation (str): Activation function
         norm (bool): Whether to use normalization
     """
 
@@ -230,15 +272,23 @@ class UpBlock(nn.Module):
         self,
         in_channels: int,
         out_channels: int,
+        has_attn: bool = False,
+        activation: str = "gelu",
         norm: bool = False,
     ):
         super().__init__()
         # The input has `in_channels + out_channels` because we concatenate the output of the same resolution
         # from the first half of the U-Net
-        self.res = ResidualBlock(in_channels + out_channels, out_channels, norm=norm)
+        self.res = ResidualBlock(in_channels + out_channels, out_channels, activation=activation, norm=norm)
+        if has_attn:
+            self.attn = AttentionBlock(out_channels)
+        else:
+            self.attn = nn.Identity()
 
     def forward(self, x: torch.Tensor):
-        return self.res(x)
+        x = self.res(x)
+        x = self.attn(x)
+        return x
 
 
 class FourierUpBlock(nn.Module):
@@ -256,6 +306,8 @@ class FourierUpBlock(nn.Module):
         out_channels: int,
         modes1: int = 16,
         modes2: int = 16,
+        has_attn: bool = False,
+        activation: str = "gelu",
         norm: bool = False,
     ):
         super().__init__()
@@ -266,11 +318,18 @@ class FourierUpBlock(nn.Module):
             out_channels,
             modes1=modes1,
             modes2=modes2,
+            activation=activation,
             norm=norm,
         )
+        if has_attn:
+            self.attn = AttentionBlock(out_channels)
+        else:
+            self.attn = nn.Identity()
 
     def forward(self, x: torch.Tensor):
-        return self.res(x)
+        x = self.res(x)
+        x = self.attn(x)
+        return x
 
 
 class MiddleBlock(nn.Module):
@@ -283,16 +342,20 @@ class MiddleBlock(nn.Module):
 
     Args:
         n_channels (int): Number of channels in the input and output.
+        has_attn (bool, optional): Whether to use attention block. Defaults to False.
+        activation (str): Activation function to use. Defaults to "gelu".
         norm (bool, optional): Whether to use normalization. Defaults to False.
     """
 
-    def __init__(self, n_channels: int, norm: bool = False):
+    def __init__(self, n_channels: int, has_attn: bool = False, activation: str = "gelu", norm: bool = False):
         super().__init__()
-        self.res1 = ResidualBlock(n_channels, n_channels, norm=norm)
-        self.res2 = ResidualBlock(n_channels, n_channels, norm=norm)
+        self.res1 = ResidualBlock(n_channels, n_channels, activation=activation, norm=norm)
+        self.attn = AttentionBlock(n_channels) if has_attn else nn.Identity()
+        self.res2 = ResidualBlock(n_channels, n_channels, activation=activation, norm=norm)
 
     def forward(self, x: torch.Tensor):
         x = self.res1(x)
+        x = self.attn(x)
         x = self.res2(x)
         return x
 
@@ -327,6 +390,146 @@ class Downsample(nn.Module):
         return self.conv(x)
 
 
+class Unet(nn.Module):
+    """Modern U-Net architecture
+
+    This is a modern U-Net architecture with wide-residual blocks and spatial attention blocks
+
+    Args:
+        in_channels (int): channels for model input
+        out_channels (int): channels for model output
+        hidden_channels (int): Number of channels in the hidden layers
+        activation (str): Activation function to use
+        norm (bool): Whether to use normalization
+        ch_mults (list): List of channel multipliers for each resolution
+        is_attn (list): List of booleans indicating whether to use attention blocks
+        mid_attn (bool): Whether to use attention block in the middle block
+        n_blocks (int): Number of residual blocks in each resolution
+        use1x1 (bool): Whether to use 1x1 convolutions in the initial and final layers
+    """
+
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        hidden_channels: int,
+        activation: str,
+        norm: bool = False,
+        ch_mults: Union[Tuple[int, ...], List[int]] = (1, 2, 2, 4),
+        is_attn: Union[Tuple[bool, ...], List[bool]] = (False, False, False, False),
+        mid_attn: bool = False,
+        n_blocks: int = 2,
+        use1x1: bool = False,
+    ) -> None:
+        super().__init__()
+        self.in_channels = in_channels
+        self.out_channels = out_channels
+        self.hidden_channels = hidden_channels
+
+        self.activation: nn.Module = ACTIVATION_REGISTRY.get(activation, None)
+        if self.activation is None:
+            raise NotImplementedError(f"Activation {activation} not implemented")
+        # Number of resolutions
+        n_resolutions = len(ch_mults)
+
+        n_channels = hidden_channels
+        # Project image into feature map
+        if use1x1:
+            self.image_proj = nn.Conv2d(in_channels, n_channels, kernel_size=1)
+        else:
+            self.image_proj = nn.Conv2d(in_channels, n_channels, kernel_size=(3, 3), padding=(1, 1))
+
+        # #### First half of U-Net - decreasing resolution
+        down = []
+        # Number of channels
+        out_channels = in_channels = n_channels
+        # For each resolution
+        for i in range(n_resolutions):
+            # Number of output channels at this resolution
+            out_channels = in_channels * ch_mults[i]
+            # Add `n_blocks`
+            for _ in range(n_blocks):
+                down.append(
+                    DownBlock(
+                        in_channels,
+                        out_channels,
+                        has_attn=is_attn[i],
+                        activation=activation,
+                        norm=norm,
+                    )
+                )
+                in_channels = out_channels
+            # Down sample at all resolutions except the last
+            if i < n_resolutions - 1:
+                down.append(Downsample(in_channels))
+
+        # Combine the set of modules
+        self.down = nn.ModuleList(down)
+
+        # Middle block
+        self.middle = MiddleBlock(out_channels, has_attn=mid_attn, activation=activation, norm=norm)
+
+        # #### Second half of U-Net - increasing resolution
+        up = []
+        # Number of channels
+        in_channels = out_channels
+        # For each resolution
+        for i in reversed(range(n_resolutions)):
+            # `n_blocks` at the same resolution
+            out_channels = in_channels
+            for _ in range(n_blocks):
+                up.append(
+                    UpBlock(
+                        in_channels,
+                        out_channels,
+                        has_attn=is_attn[i],
+                        activation=activation,
+                        norm=norm,
+                    )
+                )
+            # Final block to reduce the number of channels
+            out_channels = in_channels // ch_mults[i]
+            up.append(UpBlock(in_channels, out_channels, has_attn=is_attn[i], activation=activation, norm=norm))
+            in_channels = out_channels
+            # Up sample at all resolutions except last
+            if i > 0:
+                up.append(Upsample(in_channels))
+
+        # Combine the set of modules
+        self.up = nn.ModuleList(up)
+
+        if norm:
+            self.norm = nn.GroupNorm(8, n_channels)
+        else:
+            self.norm = nn.Identity()
+        if use1x1:
+            self.final = nn.Conv2d(in_channels, self.out_channels, kernel_size=1)
+        else:
+            self.final = nn.Conv2d(in_channels, self.out_channels, kernel_size=(3, 3), padding=(1, 1))
+
+    def forward(self, x: torch.Tensor):
+        #assert x.dim() == 5
+        orig_shape = x.shape
+        x = self.image_proj(x)
+
+        h = [x]
+        for m in self.down:
+            x = m(x)
+            h.append(x)
+
+        x = self.middle(x)
+
+        for m in self.up:
+            if isinstance(m, Upsample):
+                x = m(x)
+            else:
+                # Get the skip connection from first half of U-Net and concatenate
+                s = h.pop()
+                x = torch.cat((x, s), dim=1)
+                x = m(x)
+
+        x = self.final(self.activation(self.norm(x)))
+        return x
 
 class FourierUnet(nn.Module):
     """Unet with Fourier layers in early downsampling blocks.
@@ -342,11 +545,10 @@ class FourierUnet(nn.Module):
         n_fourier_layers (int): Number of early downsampling layers to use Fourier layers in.
         mode_scaling (bool): Whether to scale the number of modes with resolution.
     """
-
     def __init__(
         self,
-        input_channels: int,
-        output_channels: int,
+        in_channels: int,
+        out_channels: int,
         hidden_channels: int,
         modes1: int = 12,
         modes2: int = 12,
@@ -357,17 +559,16 @@ class FourierUnet(nn.Module):
         mode_scaling: bool = True,
     ) -> None:
         super().__init__()
-        self.input_channels = input_channels
-        self.output_channels = output_channels
+        self.in_channels = in_channels
+        self.out_channels = out_channels
         self.hidden_channels = hidden_channels
 
         # Number of resolutions
         n_resolutions = len(ch_mults)
 
-        insize = self.input_channels
         n_channels = hidden_channels
         # Project image into feature map
-        self.image_proj = nn.Conv2d(insize, n_channels, kernel_size=(3, 3), padding=(1, 1))
+        self.image_proj = nn.Conv2d(self.in_channels, n_channels, kernel_size=(3, 3), padding=(1, 1))
 
         # #### First half of U-Net - decreasing resolution
         down = []
@@ -441,11 +642,10 @@ class FourierUnet(nn.Module):
             self.norm = nn.GroupNorm(8, n_channels)
         else:
             self.norm = nn.Identity()
-        out_channels = self.output_channels
 
-        self.final = nn.Conv2d(in_channels, out_channels, kernel_size=(3, 3), padding=(1, 1))
+        self.final = nn.Conv2d(in_channels, self.out_channels, kernel_size=(3, 3), padding=(1, 1))
         self.activation = nn.GELU()
-        
+
     def forward(self, x: torch.Tensor):
         x = self.image_proj(x)
