+++ src/op_lib/temp_trainer.py
@@ -15,9 +15,10 @@ from .metrics import compute_metrics, write_metrics
 from .losses import LpLoss
 from .plt_util import plt_temp, plt_iter_mae
 from .heatflux import heatflux
-import .dist_utils
+from .dist_utils import local_rank, is_leader_process
 
 from torch.cuda import nvtx 
+import time
 
 t_bulk_map = {
     'wall_super_heat': 58,
@@ -44,28 +45,30 @@ class TempTrainer:
         self.val_variable = val_variable
         self.writer = writer
         self.cfg = cfg
-        self.loss = LpLoss(d=2)
+        self.loss = LpLoss(d=2, reduce_dims=[0, 1])
 
         self.push_forward_steps = push_forward_steps
         self.future_window = future_window
-        self.local_rank = dist_utils.local_rank() 
+        self.local_rank = local_rank() 
 
     def train(self, max_epochs):
         for epoch in range(max_epochs):
             print('epoch ', epoch)
             self.train_step(epoch)
             self.val_step(epoch)
+            self.lr_scheduler.step()
 
     def _forward_int(self, temp, vel):
         input = torch.cat((temp, vel), dim=1)
         pred = self.model(input)
+        return pred
         # TODO: account for different timesteps of training data
         #  - u_k+l = u_k + (t_k+l - t_k)d_l
         # at the moment, time discretization is 1, so t_k+l - t_k is just l + 1
-        timesteps = (torch.arange(self.future_window) + 1).to(pred.device).unsqueeze(-1).unsqueeze(-1)
-        last_temp_input = temp[:, -1].unsqueeze(1).to(pred.device)
-        sol = last_temp_input + timesteps * pred
-        return sol
+        #timesteps = (torch.arange(self.future_window) + 1).to(pred.device).unsqueeze(-1).unsqueeze(-1)
+        #last_temp_input = temp[:, -1].unsqueeze(1).to(pred.device)
+        #sol = last_temp_input + timesteps * pred
+        #return sol
 
     def push_forward_trick(self, temp, vel):
         #with torch.no_grad():
@@ -84,12 +87,11 @@ class TempTrainer:
             label = label.to(self.local_rank).float()
             
             pred = self.push_forward_trick(temp, vel)
-            loss = F.mse_loss(pred, label)
-            #loss = self.loss(pred, label).mean()
+            #loss = F.mse_loss(pred, label)
+            loss = self.loss(pred, label)
             self.optimizer.zero_grad()
             loss.backward()
             self.optimizer.step()
-            self.lr_scheduler.step()
             
             print(f'train loss: {loss}')
             global_iter = epoch * len(self.train_dataloader) + iter
@@ -111,13 +113,15 @@ class TempTrainer:
             write_metrics(pred, label, global_iter, 'Val', self.writer)
             del temp, vel, label
 
-    def test(self, dataset, max_timestep=100):
-        if dist_utils.is_leader_process():
+    def test(self, dataset, max_timestep=200):
+        if is_leader_process():
             self.model.eval()
             temps = []
             labels = []
-            time = min(len(dataset), max_timestep)
-            for timestep in range(0, time, self.future_window):
+            time_lim = min(len(dataset), max_timestep)
+            
+            start = time.time()
+            for timestep in range(0, time_lim, self.future_window):
                 coords, temp, vel, label = dataset[timestep]
                 temp = temp.to(self.local_rank).float().unsqueeze(0)
                 vel = vel.to(self.local_rank).float().unsqueeze(0)
@@ -128,10 +132,12 @@ class TempTrainer:
                     dataset.write_temp(temp.permute((1, 2, 0)), timestep)
                     temps.append(temp.detach().cpu())
                     labels.append(label.detach().cpu())
+            dur = time.time() - start
+            print(f'rollout time {dur} (s)')
 
             temps = torch.cat(temps, dim=0)
             labels = torch.cat(labels, dim=0)
-            dfun = dataset.get_dfun().permute((2, 0, 1))[0:time:self.future_window]
+            dfun = dataset.get_dfun().permute((2, 0, 1))[0:time_lim:self.future_window]
 
             metrics = compute_metrics(temps, labels, dfun)
             print(metrics)
